create table slots_user

+++ database/migrations/2024_04_22_172751_update_slots_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::table('slots', function (Blueprint $table) {
             $table->integer('id_city')->nullable()->after('id');
-            $table->integer('kef')->default(1)->after('kol');
+            $table->float('kef')->default(1)->after('kol');
         });
     }
